Move rendered import-excel sheet into component
