+++ app/javascript/components/table_view/style.scss
@@ -1,0 +1,11 @@
+#gridctr {
+  height: 637px;
+
+  canvas, canvas-datagrid {
+    width: 100%;
+    height: 100%;
+    --cdg-height: 100%;
+    overflowX: 'scroll';
+    overflowY: 'scroll';
+  }
+}
