+++ app/javascript/packs/table_view.js
@@ -1,0 +1,1 @@
+import 'components/table_view'
