+++ app/views/imports/show.html.erb
@@ -60,4 +60,7 @@
   <%= link_to @import.xlsx_file.filename, url_for(@import.xlsx_file) %>
 </p>
 
-<div id="gridctr" style="height: 637px;" data-url="<%= url_for @import.xlsx_file %>"></div>
+<div id="gridctr" data-url="<%= url_for @import.xlsx_file %>"></div>
+
+<%= javascript_pack_tag 'table_view' %>
+<%= stylesheet_pack_tag 'table_view' %>
